new seeds text was added
